Build the UserAssertion exchange the provided token for a token we can call the graph with.
These steps sort of tell a story of discovery of the non-obvious steps (at least they were to me) required to get things working. Each step is going to build on the last, getting closer to the final product.

At this point, building the UserAssertion will fail because the bootstrap context is null (the signin key wasn't preserved)