Save the signin token so we can use the bootstrap context to create a user assertion.
at this point, the token will still fail to acquire because the user has never given consent for the webapi to access the graph on its behalf. normally this would be done at login time, but we already logged in for the other app. We can, however, get a URL that we can have the user go to to give consent (with GetAuthorizationRequestUrlAsync). We can flow that URL back to the client webapp, which can send the user to the page, get consent, and then we are all set!