get the AuthorizationUrl for the user to grant consent with.
In order to grant consent to the WebApi, you have to get an AuthorizationUrl. In order to get an AuthorizationUrl, you have to have a redirectUri. When you register a WebApi, you don't have an option to create a redirectUri. So, you have to go to your WebApi registration and ADD a Web platform with a redirectUri specifically to allow consent to be granted. (This doesn't seem like a great way to do it, but I'm getting desperate and can't figure out another way).  Also, since you are going to tell the ConfidentialClientApplication that it should redirect to a particular URI, we need to make sure there's really a page there to redirect to.

At this point, when you call the API, you will see the error returned with the URL value. If you copy/paste that URL and go to it, you will be prompted to login, and grant consent. Once you do that, all subsequent calls to the WebApi will succeed.

NOTE: This current incarnation uses a nasty hack to return a hyperlink to the WebApp to click on. This is pretty dirty and the next iteration will remove that with a more formal handshare between WebApp and WebApi