Implement WebApp side of consent granting.
This will just present the user with an error message and a link to click to grant consent. We *could* automatically redirect to the link which would ask for consent (which might be jarring for the user). But then how to redirect back? We *could* pass another redirect parameter (in the state) so the auth.aspx page for the WebApi would redirect back to our WebApp, but then we still have to find a way to get the user to understand that they have to retry the operation.

For now, the point of this exercise is just to show how to get the url, how to flow the consent, and (in the next iterations), how to call the MS Graph in the WebApi on the user's behalf.  The whole redirecting thing is an exercise for the reader.

(or hopefully someone will point out a better consent workflow!!)